refactor: move all photo endpoints into own folders/files
