+++ server/src/routes/photos/deletePhoto.ts
