+++ server/src/routes/places/types.ts
@@ -31,12 +31,3 @@ export type PlaceDetails = Place & {
   userRating: number;
   creatorUsername: string;
 };
-
-export type DatabasePhoto = {
-  photo_file_key: string;
-  place_id: PlaceId;
-  user_id: string;
-  upload_timestamp: Date;
-};
-
-export type Photo = { fileKey: string; base64Image: string; userId: string };
